server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
]
